+++ app.js
@@ -6,7 +6,6 @@ const logger = require('morgan')
 const cors = require('cors')
 const jwt = require('jsonwebtoken')
 require('dotenv').config()
-const multer = require("multer");
 
 const indexRouter = require('./routes/index')
 
@@ -21,30 +20,6 @@ app.use(express.json())
 app.use(express.urlencoded({ extended: false }))
 app.use(cookieParser())
 app.use(express.static(path.join(__dirname, 'public')))
-
-
-const storage = multer.diskStorage({
-  destination: path.join(__dirname, "public/images"),
-  
-  filename:(req, file, cb)=>{
-    let imageName = Date.now() + "-" + path.extname(file.originalname);
-    cb(null, imageName)
-  }
-})
-
-app.use(multer({
-  storage,
-   dest: path.join(__dirname, "public/images"),
-   fileFilter:(req,file,cb)=>{
-    const filetypes = /image|png|jpg|svg|webp/;
-    const mimetype = filetypes.test(file.mimetype);
-    const extname = filetypes.test(path.extname(file.originalname))
-    if(mimetype && extname){
-      return cb(null, true)
-    }
-    cb("Debe ser una imagen valida")
-   }
-}).single("image"))
 
 
 app.use('/', indexRouter)
