+++ routes/users.js
@@ -1,7 +1,7 @@
 
 const express = require('express');
 const { deleteU, create, update, list,detail, servicioimagenpost} = require('../controllers/users')
-const { validateCreate, validateToken } = require('../middlewares/index')
+const { validateCreate, validateToken, imagen } = require('../middlewares/index')
 
 const router = express.Router();
 
@@ -11,6 +11,6 @@ router.post("/create",validateCreate,create)
 router.put("/update/:id",update)
 router.delete("/delete/:id",deleteU)
 
-router.post("/imagen" ,servicioimagenpost);
+router.post("/imagen", imagen,servicioimagenpost);
 
 module.exports = router
